alles werkt behalve de datum
de where bij revenueDAO moet nog worden aangepast en bij de ik krijg de datum niet goed terug

+++ SomerenModel/Revenue.cs
@@ -9,7 +9,7 @@ namespace SomerenModel
     public class Revenue
     {
         public int Sales { get; set; }
-        public float Ternover { get; set; }
+        public double Ternover { get; set; }
         public int NumberOfCustomers { get; set; }
     }
 }
